non compile backup default impls are close for coverage

+++ src/org/adligo/tests4j/models/shared/coverage/I_LineCoverage.java
@@ -66,7 +66,7 @@ public interface I_LineCoverage {
 	 * @param p
 	 * @return
 	 */
-	public I_LineCoverageSegment getLineCoverageSegment(int p);
+	public I_LineCoverageSegment getLineCoverageSegment(short p);
 	
 
 }
